:sparkles: Scene实现model构建缓存以及相关类的哈希算法
:bug: 修复内建hash对str类型返回值随机化的问题（TODO: 引入新的确定性哈希方式）
:bug: 修复内建Prism生成PolyData时数据类型非float的警告
:heavy_plus_sign: 添加cloudpickle依赖

+++ metalpy/scab/modelling/shapes/prism.py
@@ -131,5 +131,5 @@ class Prism(Shape3D):
 
         faces = np.r_[top_face, bottom_face, side_faces]
 
-        shape = pv.PolyData(vertices, faces=faces)
+        shape = pv.PolyData(vertices.astype(np.float64), faces=faces)
         return shape
